pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "vau"
version = "0.1.0"
description = ""
authors = ["Niklas Rosenstein <rosensteinniklas@gmail.com>"]
license = "MIT"
readme = "readme.md"
packages = [{ include = "vau", from = "src" }]
classifiers = []
keywords = []

[tool.poetry.urls]
# "Bug Tracker" = ""
# Documentation = ""
# Homepage = ""
# Repository = ""

[tool.poetry.dependencies]
python = "^3.6"
typer = "^0.7.0"
hvac = "^1.1.0"
rich = "^13.3.4"
"databind.json" = "^4.2.4"
typer-builder = "^0.0.8"

[tool.poetry.dev-dependencies]
black = "*"
flake8 = "*"
isort = "*"
mypy = "*"
pytest = "*"

[tool.poetry.scripts]
vau = "vau.__main__:main"

[tool.slap]
typed = true

[tool.slap.test]
check = "slap check"
mypy = "dmypy run src/"
pytest = "pytest tests/ -vv"
black = "black --check src/ tests/"
isort = "isort --check-only src/ tests/"
flake8 = "flake8 src/ tests/"

[tool.slap.run]
fmt = "black src/ tests/ && isort src/ tests/"

[tool.mypy]
python_version = "3.6"
explicit_package_bases = true
mypy_path = ["src"]
namespace_packages = true
pretty = true
show_error_codes = true
show_error_context = true
strict = true
warn_no_return = true
warn_redundant_casts = true
warn_unreachable = true
warn_unused_ignores = true

[tool.isort]
profile = "black"
line_length = 120
combine_as_imports = true

[tool.black]
line-length = 120
